+++ core/exec/test/src/mill/exec/CrossTests.scala
@@ -1,20 +1,75 @@
 package mill.exec
 
-import mill.testkit.UnitTester
+import mill.define.Discover
+import mill.{Cross, Task}
+import mill.testkit.{TestBaseModule, UnitTester}
 import mill.testkit.UnitTester.Result
 import mill.util.TestGraphs
-import mill.util.TestGraphs.{
-  crossResolved,
-  doubleCross,
-  nestedCrosses,
-  innerCrossModule,
-  singleCross,
-  nonStringCross,
-  crossExtension
-}
-import utest._
+import mill.util.TestGraphs.{crossResolved, doubleCross, nestedCrosses, nonStringCross, singleCross}
+import utest.*
 
 object CrossTests extends TestSuite {
+  
+  object crossExtension extends TestBaseModule {
+    object myCross extends Cross[MyCrossModule]("a", "b")
+    trait MyCrossModule extends Cross.Module[String] {
+      def param1 = Task { "Param Value: " + crossValue }
+    }
+
+    object myCrossExtended extends Cross[MyCrossModuleExtended](("a", 1), ("b", 2))
+    trait MyCrossModuleExtended extends MyCrossModule with Cross.Module2[String, Int] {
+      def param2 = Task { "Param Value: " + crossValue2 }
+    }
+
+    object myCrossExtendedAgain
+        extends Cross[MyCrossModuleExtendedAgain](("a", 1, true), ("b", 2, false))
+    trait MyCrossModuleExtendedAgain extends MyCrossModuleExtended
+        with Cross.Module3[String, Int, Boolean] {
+      def param3 = Task { "Param Value: " + crossValue3 }
+    }
+    lazy val millDiscover = Discover[this.type]
+  }
+
+  object innerCrossModule extends TestBaseModule {
+    object myCross extends Cross[MyCrossModule]("a", "b")
+    trait MyCrossModule extends Cross.Module[String] {
+      object foo extends CrossValue {
+        def bar = Task { "foo " + crossValue }
+      }
+
+      object baz extends CrossValue {
+        def bar = Task { "baz " + crossValue }
+      }
+    }
+
+    object myCross2 extends Cross[MyCrossModule2](("a", 1), ("b", 2))
+    trait MyCrossModule2 extends Cross.Module2[String, Int] {
+      object foo extends InnerCrossModule2 {
+        def bar = Task { "foo " + crossValue }
+        def qux = Task { "foo " + crossValue2 }
+      }
+      object baz extends InnerCrossModule2 {
+        def bar = Task { "baz " + crossValue }
+        def qux = Task { "baz " + crossValue2 }
+      }
+    }
+
+    object myCross3 extends Cross[MyCrossModule3](("a", 1, true), ("b", 2, false))
+    trait MyCrossModule3 extends Cross.Module3[String, Int, Boolean] {
+      object foo extends InnerCrossModule3 {
+        def bar = Task { "foo " + crossValue }
+        def qux = Task { "foo " + crossValue2 }
+        def lol = Task { "foo " + crossValue3 }
+      }
+      object baz extends InnerCrossModule3 {
+        def bar = Task { "baz " + crossValue }
+        def qux = Task { "baz " + crossValue2 }
+        def lol = Task { "baz " + crossValue3 }
+      }
+    }
+    lazy val millDiscover = Discover[this.type]
+  }
+  
   val tests = Tests {
     test("singleCross") {
       val check = UnitTester(singleCross, null)
